Stage 2 + G7 review fixes (output-side per-slot correction, LoRA-path test, KV-cached eval, resume safety)
Both review passes landed before the Fable quota ended them mid-polish; their
substantive edits are here and verified (lint clean, 76 tests green):
- gist_vocab: per-slot rank-64 correction on the OUTPUT side too (P_out/Q_out),
  resolving sonnet's flagged judgement call; param count pinned at ~37M.
- run_g7: test proving gradients reach BOTH the g7 LoRA and GistVocab through
  pm.get_decoder() (if that path bypassed PEFT the LoRA would never train),
  gradient checkpointing, KV-cached greedy group prediction matching uncached,
  checkpoint roundtrip/resume, slot-7 confusion split, mu-dict load guards.
- stage 2: contamination guard widened to all 1319 GSM8K-test questions (a
  strict superset of the eval sets), partial trailing shard never counted
  complete, resume loses no docs, hard-split pieces verified within max_span
  after decode inflation.

Review seat note: Fable is out of quota; Opus verified the remaining
GPU-hazard items (RAM clauses in GB, encode settings shared with stage 1,
eval ids reproduced from the same candidate window).

+++ src/marker/g7.py
@@ -169,7 +169,9 @@ def two_gather_loss(
     n_text = int(text_mask.sum())
     text_ce = zero
     if n_text:
-        text_logits = head.base_head(hidden[text_mask])
+        # .float(): CE in bf16 over a 150k-wide (text) or 4096-wide (gist)
+        # softmax adds ~1e-2 noise -- cheap to do the log-softmax in fp32.
+        text_logits = head.base_head(hidden[text_mask]).float()
         text_ce = func.cross_entropy(text_logits, targets[text_mask])
 
     n_gist = int(gist_mask.sum())
@@ -185,7 +187,7 @@ def two_gather_loss(
             sel = slots == s
             block_rows = rows[s * vocab.K : (s + 1) * vocab.K].to(gist_hidden.dtype)
             block_bias = bias[s * vocab.K : (s + 1) * vocab.K].to(gist_hidden.dtype)
-            block_logits = gist_hidden[sel] @ block_rows.T + block_bias
+            block_logits = (gist_hidden[sel] @ block_rows.T + block_bias).float()
             local_target = gist_targets[sel] - (vocab.base_vocab + s * vocab.K)
             losses.append(func.cross_entropy(block_logits, local_target, reduction="sum"))
         gist_ce = torch.stack(losses).sum() / n_gist
